Refactored to move clients to specific report client folder
