+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/client/FdcFilesClient.java
@@ -1,10 +1,9 @@
-package uk.gov.justice.laa.crime.dces.report.maatapi.client;
+package uk.gov.justice.laa.crime.dces.report.client;
 
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.service.annotation.GetExchange;
 import uk.gov.justice.laa.crime.dces.report.model.ContributionFilesResponse;
-import uk.gov.justice.laa.crime.dces.report.service.ContributionFilesClient;
 
 import java.time.LocalDate;
